Rename vm-ssh-config to local-ssh-config
